+++ src/scope/mod.rs
@@ -4,7 +4,7 @@ use crate::ModulePath;
 mod display;
 
 #[derive(Debug, Default)]
-pub struct CodeSection {
+pub struct Scope {
     pub(crate) functions: Functions,
     pub(crate) required_modules: Vec<ModulePath>,
     pub(crate) variables: Variables,