+++ src/source_code/mod.rs
@@ -28,9 +28,9 @@ pub type CharOffset = usize;
 pub struct SourceCode {
     pub filename: path::PathBuf,
     pub(crate) lines: LineNumber,
-    pub(crate) length_of_code_section: LineNumber,
+    pub(crate) length_of_scope: LineNumber,
     pub(crate) length_of_csv_section: LineNumber,
-    pub(crate) code_section: Option<String>,
+    pub(crate) scope: Option<String>,
     pub(crate) csv_section: String,
     pub(crate) original: String,
 }
@@ -50,17 +50,17 @@ impl SourceCode {
     pub fn new<S: Into<String>>(input: S, filename: path::PathBuf) -> Result<SourceCode> {
         let str_input: String = input.into();
 
-        if let Some((code_section, csv_section)) = str_input.split_once(CODE_SECTION_SEPARATOR) {
+        if let Some((scope, csv_section)) = str_input.split_once(CODE_SECTION_SEPARATOR) {
             let csv_lines = csv_section.lines().count();
-            let code_lines = code_section.lines().count();
+            let code_lines = scope.lines().count();
 
             Ok(SourceCode {
                 filename,
                 lines: csv_lines + code_lines,
-                length_of_code_section: code_lines,
+                length_of_scope: code_lines,
                 length_of_csv_section: csv_lines,
                 csv_section: csv_section.to_string(),
-                code_section: Some(code_section.to_string()),
+                scope: Some(scope.to_string()),
                 original: str_input.to_owned(),
             })
         } else {
@@ -69,10 +69,10 @@ impl SourceCode {
             Ok(SourceCode {
                 filename,
                 lines: csv_lines,
-                length_of_code_section: 0,
+                length_of_scope: 0,
                 length_of_csv_section: csv_lines,
                 csv_section: str_input.to_owned(),
-                code_section: None,
+                scope: None,
                 original: str_input.to_owned(),
             })
         }
@@ -96,7 +96,7 @@ impl SourceCode {
 
     pub(crate) fn csv_line_number(&self, position: a1_notation::Address) -> LineNumber {
         let row = position.row.y;
-        self.length_of_code_section + 1 + row
+        self.length_of_scope + 1 + row
     }
 
     pub(crate) fn line_offset_for_cell(&self, position: a1_notation::Address) -> CharOffset {
@@ -139,9 +139,9 @@ mod tests {
         SourceCode {
             filename: path::PathBuf::from("test.csvpp".to_string()),
             lines: 25,
-            length_of_code_section: 10,
+            length_of_scope: 10,
             length_of_csv_section: 15,
-            code_section: Some("\n".repeat(10)),
+            scope: Some("\n".repeat(10)),
             csv_section: "foo,bar,baz".to_string(),
             original: "\n\n\n\n\n\n\n\n\n\n---\nfoo,bar,baz".to_string(),
         }
@@ -237,19 +237,19 @@ foo1,bar1,baz1
     }
 
     #[test]
-    fn open_no_code_section() {
+    fn open_no_scope() {
         let source_code =
             SourceCode::new("foo,bar,baz", std::path::PathBuf::from("foo.csvpp")).unwrap();
 
         assert_eq!(source_code.lines, 1);
         assert_eq!(source_code.length_of_csv_section, 1);
-        assert_eq!(source_code.length_of_code_section, 0);
-        assert_eq!(source_code.code_section, None);
+        assert_eq!(source_code.length_of_scope, 0);
+        assert_eq!(source_code.scope, None);
         assert_eq!(source_code.csv_section, "foo,bar,baz".to_string());
     }
 
     #[test]
-    fn open_code_section() {
+    fn open_scope() {
         let s = TestSourceCode::new(
             "csv",
             "
@@ -263,8 +263,8 @@ foo,bar,baz,=foo
 
         assert_eq!(source_code.lines, 5);
         assert_eq!(source_code.length_of_csv_section, 2);
-        assert_eq!(source_code.length_of_code_section, 3);
-        assert_eq!(source_code.code_section, Some("\nfoo := 1\n\n".to_string()));
+        assert_eq!(source_code.length_of_scope, 3);
+        assert_eq!(source_code.scope, Some("\nfoo := 1\n\n".to_string()));
         assert_eq!(source_code.csv_section, "\nfoo,bar,baz,=foo\n".to_string());
     }
 }